Move classes into components package instead of buttons package
This will allow for a better future package layout
